requirements.txt
asgiref==3.2.7
asn1crypto==0.24.0
astroid==2.2.5
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
colorama==0.4.1
cotidia-import-export==0.1.5
cryptography==2.7
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.2.6
django-admin-rangefilter==0.5.3
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap3==11.0.0
django-bootstrap4-datetimepicker==4.2
django-channels==0.7.0
django-ckeditor==5.9.0
django-crispy-forms==1.8.1
django-dbbackup==3.3.0
django-dbbackup-ui==0.2.1
django-filter==2.1.0
django-import-export==2.0
django-js-asset==1.2.2
django-money==0.15
django-money-rates==0.3.1
django-mysql==3.5.0
django-recommend==0.1.dev9
django-scheduler==0.9.3
django-sitetree==1.14.0
django-summernote==0.8.11.4
django-taggit==1.1.0
djangorestframework==3.11.0
et-xmlfile==1.0.1
icalendar==4.0.4
idna==2.8
isort==4.3.17
jdcal==1.4.1
lazy-object-proxy==1.3.1
MarkupPy==1.14
mccabe==0.6.1
mysqlclient==1.4.6
nexmo==2.4.0
numpy==1.17.4
oauthlib==3.1.0
odfpy==1.4.0
openpyxl==3.0.2
pandas==0.25.3
Pillow==6.0.0
py-moneyed==0.8.0
pycparser==2.19
PyJWT==1.7.1
pylint==2.3.1
pyrecommend==0.1.dev5
PySocks==1.7.0
python-dateutil==2.8.1
pytz==2019.1
PyYAML==5.1.2
requests==2.22.0
requests-oauthlib==1.3.0
six==1.13.0
sqlparse==0.3.0
tablib==0.12.1
twilio==6.28.0
typed-ast==1.3.1
unicodecsv==0.14.1
urllib3==1.25.3
whitenoise==4.1.2
wrapt==1.11.1
xlrd==1.2.0
xlwt==1.3.0
